kvm_rs::vcpu
Struct Vcpu

pub struct Vcpu { /* private fields */ }

Wrapper for VCPU ioctls.

Representation of the file descriptor obtained by the KVM_CREATE_VCPU ioctl. This wrapper provides access to the VCPU ioctls as described in KVM API.

Implementations

impl Vcpu

pub fn get_regs(&self) -> Result<kvm_regs>

Get the general purpose registers with the KVM_GET_REGS ioctl in form of kvm_regs.

pub fn set_regs(&self, regs: kvm_regs) -> Result<()>

Set the general purpose registers with the KVM_SET_REGS ioctl in form of kvm_regs.

pub fn get_sregs(&self) -> Result<kvm_sregs>

Get the special registers with the KVM_GET_SREGS ioctl in form of kvm_sregs.

pub fn set_sregs(&self, sregs: kvm_sregs) -> Result<()>

Set the special registers with the KVM_SET_SREGS ioctl in form of kvm_sregs.

pub fn get_debugregs(&self) -> Result<kvm_debugregs>

Get the debug registers with the KVM_GET_DEBUGREGS ioctl in form of kvm_debugregs.

pub fn set_debugregs(&self, dregs: kvm_debugregs) -> Result<()>

Set the debug registers with the KVM_SET_DEBUGREGS ioctl in form of kvm_debugregs.

pub fn set_single_step(&self, enable: bool) -> Result<()>

Enable or disable guest single steppig (debug) with the KVM_GUESTDBG_ENABLE ioctl.

pub fn run(&mut self) -> Result<KvmExit<'_>>

Run the guest VCPU with the KVM_RUN ioctl until it exits with one of the exit reasons described in KvmExit.